delete fix

+++ src/main/java/mif/vu/ikeea/RepositoryService/UserService.java
@@ -20,15 +20,6 @@ public class UserService implements UserDetailsService {
     @Autowired
     UserRepository userRepository;
 
-    @Autowired
-    private LearningDayService learningDayService;
-
-    @Autowired
-    private GoalService goalService;
-
-    @Autowired
-    private CommentService commentService;
-
     @Transactional
     public ApplicationUser add(ApplicationUser user) {
         return userRepository.save(user);
@@ -36,7 +27,12 @@ public class UserService implements UserDetailsService {
 
     @Transactional
     public void delete(Long id) {
-        deleteUserData(id);
+        ApplicationUser user = loadById(id);
+        if(!user.getChildren().isEmpty()) {
+            throw new UserDeleteException("Leader can't be deleted.");
+        }
+        ApplicationUser manager = user.getManager();
+        manager.getChildren().remove(user);
         userRepository.deleteById(id);
     }
 
@@ -137,32 +133,5 @@ public class UserService implements UserDetailsService {
             authorities.add(new SimpleGrantedAuthority("ROLE_" + role.getName()));
         });
         return authorities;
-    }
-
-    public void deleteUserData(Long id) {
-        ApplicationUser user = loadById(id);
-        if(!user.getChildren().isEmpty()) {
-            throw new UserDeleteException("Leader can't be deleted.");
-        }
-        List<LearningDay> learningDays = learningDayService.getAllByUserId(id);
-        if(!learningDays.isEmpty()) {
-            for (LearningDay learningDay : learningDays) {
-                learningDayService.delete(learningDay.getId());
-            }
-        }
-        List<Goal> goals = goalService.getAllByUserId(id);
-        if(!goals.isEmpty()) {
-            for (Goal goal : goals) {
-                goalService.delete(goal.getId());
-            }
-        }
-        List<Comment> comments = commentService.getAllByUserId(id);
-        if(!comments.isEmpty()) {
-            for (Comment comment : comments) {
-                commentService.delete(comment.getId());
-            }
-        }
-        ApplicationUser manager = user.getManager();
-        manager.getChildren().remove(user);
     }
 }
